Conflicts:
	nosolorice/src/main/webapp/WEB-INF/temple/header/header.jsp

+++ nosolorice/src/main/java/com/nosolorice/app/hyuncontroller/AdminController.java
@@ -40,7 +40,7 @@ public class AdminController {
 	
 	@RequestMapping("joinApproveProcess.ajax")
 	@ResponseBody
-	public Map<String, String> joinApproveProcess(int businessNumber, int no){	
+	public Map<String, String> joinApproveProcess(long businessNumber, int no){	
 		// 사업자 회원정보의 상태값 y로 바꾸기
 		adminService.updateJoinApprove(businessNumber);
 		adminService.deleteJoinApprove(no);
